SearchService-02

+++ pkg/yarus_platform/data/domain/entity/search_service.go
@@ -9,5 +9,5 @@ import (
 type SearchService interface {
 	First(ctx context.Context, cond *selection_condition.SelectionCondition, langID uint) (*Entity, error)
 	Query(ctx context.Context, cond *selection_condition.SelectionCondition, langID uint) ([]Entity, error)
-	Count(ctx context.Context, cond *selection_condition.SelectionCondition) (int64, error)
+	Count(ctx context.Context, cond *selection_condition.SelectionCondition) (uint, error)
 }
